requirements.txt
aiogram==3.17
python-dotenv==1.0.0
apscheduler==3.10.4
alembic==1.13.1